// generated/api-client/models/project-detail-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { ProjectOwnerDto } from './project-owner-dto';
// May contain unused imports in some cases
// @ts-ignore
import { ProjectRoleDto } from './project-role-dto';

/**
 * 
 * @export
 * @interface ProjectDetailDto
 */
export interface ProjectDetailDto {
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'description': string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'status': ProjectDetailDtoStatusEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof ProjectDetailDto
     */
    'tags': Array<string>;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'dueDate'?: string;
    /**
     * 
     * @type {ProjectOwnerDto}
     * @memberof ProjectDetailDto
     */
    'owner': ProjectOwnerDto;
    /**
     * 
     * @type {Array<ProjectRoleDto>}
     * @memberof ProjectDetailDto
     */
    'roles': Array<ProjectRoleDto>;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDetailDto
     */
    'updatedAt': string;
}

export const ProjectDetailDtoStatusEnum = {
    OPEN: 'OPEN',
    IN_PROGRESS: 'IN_PROGRESS',
    COMPLETED: 'COMPLETED',
    CANCELLED: 'CANCELLED'
} as const;

export type ProjectDetailDtoStatusEnum = typeof ProjectDetailDtoStatusEnum[keyof typeof ProjectDetailDtoStatusEnum];